RFC: Splits callback tests into seperate files
